outsourced LineAttachmentUtil and LineIntersect
